Added new Dev Log + tweaked basic settings
Tweaked basic settings, added IAP Receipt Validator package and a couple of basic settings in the CBS package.

Added a new Dev Log which is inside this folder but there's also a public development log named 1. Space Guardians Working Log - Gareth

This file is to be released to the public so should not hold any sensitive data.